pyproject.toml
[project]
name = "form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Easton Potokar", email = "contagon6@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "evalio>=0.4",
    "gradio-rerun>=0.24.0",
    "ipython>=9.5.0",
    "polars>=1.30.0",
    "pyvista>=0.46.3",
    "rerun-sdk>=0.23",
    "seaborn>=0.13.2",
]

[tool.scikit-build]
minimum-version = "build-system.requires"
build-dir = "build/{wheel_tag}"

[build-system]
requires = ["scikit-build-core>=0.10", "nanobind", "evalio>=0.4"]
build-backend = "scikit_build_core.build"

[dependency-groups]
dev = ["snakeviz>=2.2.2"]

[tool.scikit-build.cmake.define]
FORM_BUILD_PYTHON = true
CMAKE_BUILD_TYPE = "Release"
# CMAKE_BUILD_TYPE = "RelWithDebInfo"

[tool.uv.sources]
evalio = { path = "../evalio" }

[tool.basedpyright]
typeCheckingMode = "standard"
include = ["python"]
